pyproject.toml
[project]
name = "msai-setup"
version = "0.1.0"
description = "MS-S1 MAX Home Server Documentation and Setup Scripts"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "typer>=0.21.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
msai = "msai_setup.cli:app"

[dependency-groups]
dev = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.7.0",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-glightbox>=0.4.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pyright>=1.1.0",
    "pytest>=8.0.0",
    "types-pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/msai_setup"]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
fixable = ["ALL"]
select = ["E", "W", "F", "I", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]
"**/__init__.py" = ["D104"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.14"
typeCheckingMode = "strict"

[tool.pytest.ini_options]
testpaths = ["tests"]
